tests/wpt/web-platform-tests/referrer-policy/css-integration/font-face/internal-import-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - Font from imported stylesheet (internal)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body>
    <p>Check that resources from stylesheets (imported from internal
    stylesheets) are loaded with the referrer and referrer policy from from the
    imported style sheet.</p>

    <div class="styled"></div>

    <script>
      let css_test = async_test("Font from imported stylesheet (internal).");
      let id = token();
      let url_prefix = location.protocol + "//www1." + location.hostname + ":" +
                       location.port + "/referrer-policy/generic/subresource/";
      let css_url = url_prefix + "stylesheet.py?id=" + id + "&type=font";
      let font_url = url_prefix + "font.py?report-headers&id=" + id;

      let style = document.createElement("style");
      style.textContent = "@import url('" + css_url + "');";
      document.head.appendChild(style);
      css_test.step_timeout(
          queryXhr.bind(this, font_url,
                        function(message) {
                          css_test.step(function() {
                            assert_own_property(message, "headers");
                            assert_own_property(message, "referrer");
                            assert_equals(message.referrer, css_url);
                          });
                          css_test.done();
                        }),
          1000);
      </script>

    <div id="log"></div>
  </body>
</html>
